Fixed errs, successful tests
Errors from last night appeared this morning, so I must have not saved last night. In this version, all tests run, and the website works, showing a list of bikes and one bike when it's clicked on. Also, I had to solve a problem with a master lock file before this push.